+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/multithreading/UnsynchronizedStaticDateFormatterRule.java
@@ -4,17 +4,7 @@
 
 package net.sourceforge.pmd.lang.java.rule.multithreading;
 
-import java.util.Set;
-
-import net.sourceforge.pmd.lang.ast.Node;
-import net.sourceforge.pmd.lang.java.ast.ASTClassOrInterfaceType;
-import net.sourceforge.pmd.lang.java.ast.ASTFieldDeclaration;
-import net.sourceforge.pmd.lang.java.ast.ASTMethodDeclaration;
-import net.sourceforge.pmd.lang.java.ast.ASTSynchronizedStatement;
-import net.sourceforge.pmd.lang.java.ast.ASTVariableDeclaratorId;
-import net.sourceforge.pmd.lang.java.rule.AbstractJavaRule;
-import net.sourceforge.pmd.lang.symboltable.NameOccurrence;
-import net.sourceforge.pmd.util.CollectionUtil;
+import java.text.DateFormat;
 
 /**
  * Using a DateFormatter (SimpleDateFormatter) which is static can cause
@@ -29,37 +19,12 @@ import net.sourceforge.pmd.util.CollectionUtil;
  * 
  * @author Allan Caplan
  * @see <a href="https://sourceforge.net/p/pmd/feature-requests/226/">feature #226 Check for SimpleDateFormat as singleton?</a>
+ * @deprecated This rule is being replaced by {@link UnsynchronizedStaticFormatterRule}. The rule will be removed with PMD 7.0.0.
  */
-public class UnsynchronizedStaticDateFormatterRule extends AbstractJavaRule {
+@Deprecated
+public class UnsynchronizedStaticDateFormatterRule extends UnsynchronizedStaticFormatterRule {
 
-    private static Set<String> targets = CollectionUtil.asSet(
-            new String[] { "DateFormat", "SimpleDateFormat", "java.text.DateFormat", "java.text.SimpleDateFormat" });
-
-    @Override
-    public Object visit(ASTFieldDeclaration node, Object data) {
-        if (!node.isStatic()) {
-            return data;
-        }
-        ASTClassOrInterfaceType cit = node.getFirstDescendantOfType(ASTClassOrInterfaceType.class);
-        if (cit == null || !targets.contains(cit.getImage())) {
-            return data;
-        }
-        ASTVariableDeclaratorId var = node.getFirstDescendantOfType(ASTVariableDeclaratorId.class);
-        for (NameOccurrence occ : var.getUsages()) {
-            Node n = occ.getLocation();
-            if (n.getFirstParentOfType(ASTSynchronizedStatement.class) != null) {
-                continue;
-            }
-            // ignore usages, that don't call a method.
-            if (!n.getImage().contains(".")) {
-                continue;
-            }
-
-            ASTMethodDeclaration method = n.getFirstParentOfType(ASTMethodDeclaration.class);
-            if (method != null && !method.isSynchronized()) {
-                addViolation(data, n);
-            }
-        }
-        return data;
+    public UnsynchronizedStaticDateFormatterRule() {
+        super(DateFormat.class);
     }
 }